Vectors, Points, Colors
Add support for Vector3, Point3 and Color structs.  These are
implemented similarly to Ray Tracing In One Weekend Chapter 3 except
using Rust traits to support operations.  Implementing these for
Chapter 3 does not change the generated image in any way.

Unlike the reference book the Vector3, Point3, and Color structs are
distinct types.  There are several reasons for this.

1) Distinct types means that the operations on vectors and points can
be limited to those that make sense.  For example `P + V -> P`,
`P + P -> undefined`, `V * V -> undefined`.  Getting these correct
means catching math mistakes at compile time.

2) The Neg and Sub traits on Color are not physically based (outside
of quantum effects) and trying to use them generally means there is a
computation error.  They are not present in this implementation to
prevent that from happening, although they could be added back for
artistic reasons.

3) For physically based rendering the Color struct should be samples
in the frequency domain and not RGB which is a perceptual domain.
These are close but using RGB can lead to subtle blending issues.
This is not yet implemented but is anticipated for a far future
release.

There are some other design differences as well.

1) The operation traits do not support borrow versions at this time
and the structs support Copy.  Testing showed that #[inline] and Copy
were about the same speed as explicitly borrowing everywhere while
making the math operations less clunky.

2) Casts of Color/Point3/Vector3 to/from slices is not yet supported.
Realistically these operations will ultimately be needed in some way
to import triangle meshes efficiently.