Adding a Sphere
Implement a simple Sphere intersection test following the example in
Ray Tracing In One Weekend Chapter 5.  For simplicity the hit_sphere
function was placed in main.rs as it will subsequently be moved around
with the creation of a Hittable struct.  The code also passes the Ray
around by borrow rather than ownership as it will get reused more and
more.