Multiple Objects
Implement multiple object support for the remainder of Ray Tracing In
One Weekend Chapter 6.  For Rust this is done by creating a Hittable
trait along with a HittableList struct for handling multiple objects.
The hit function is then dynamically dispatched based upon the
underlying type.

To support parallelism the hittable objects are all wrapped with Arc
for thread safe reference counting and the dispatch type is augmented
with 'Sync + Send'.

Running this version should result in images/image-0.6-list.pnm on the
command line.