Diffuse Materials
Add Lambertian light scattering as explained in "Ray Tracing In One
Weekend" Chapter 8.  Use the corrected Lambertian formula with
random_in_hemisphere.  The other two methods from the book are
provided in the source code (commented out) for experimentation
purposes.

Also hard code a gamma of 2.0 when converting Color to u8 (in
color.rs).

The sample image is not readily reproducible because the rays are
randomly generated resulting in pixel noise that varies from run to
run.

images/image-0.8-lambertian.pnm